requirements.txt
# Core FastAPI
fastapi
uvicorn[standard]
pydantic
python-multipart
python-dotenv

# OpenAI & AI Libraries
openai
langchain
langchain-openai
langchain-community

# CrewAI for Agents
crewai
crewai-tools

# HTTP Clients
httpx
requests

# Additional Agent Dependencies
tiktoken
chromadb

# Utilities
python-dotenv

# Build Tools
setuptools
wheel
pytest
pytest-asyncio
pytest-cov
